# swagger.yaml
...
swagger: "2.0"
info:
  description: |
    BaaSapi API is an HTTP API served by BaaSapi. It is used by the BaaSapi UI and everything you can do with the UI can be done using the HTTP API.

    # Authentication

    Most of the API baask8ss require to be authenticated as well as some level of authorization to be used.
    BaaSapi API uses JSON Web Token to manage authentication and thus requires you to provide a token in the **Authorization** header of each request
    with the **Bearer** authentication mechanism.

    Example:
    ```
    Bearer eyJ...1YEE
    ```

    # Security

    Each API baask8s has an associated access policy, it is documented in the description of each baask8s.

    Different access policies are available:
    * Public access
    * Authenticated access
    * Restricted access
    * Administrator access

    ### Public access

    No authentication is required to access the baask8ss with this access policy.

    ### Authenticated access

    Authentication is required to access the baask8ss with this access policy.

    ### Restricted access

    Authentication is required to access the baask8ss with this access policy.
    Extra-checks might be added to ensure access to the resource is granted. Returned data might also be filtered.

    ### Administrator access

    Authentication as well as an administrator role are required to access the baask8ss with this access policy.

  version: "1.20.2"
  title: "BaaSapi API"
  contact:
...
host: "baasapi.domain"
basePath: "/api"
...
schemes:
...
paths:
...
  /baask8s_groups/{id}:
    get:
      tags:
      - "baask8s_groups"
      summary: "Inspect an baask8s group"
      description: |
        Retrieve details abount an baask8s group.
        **Access policy**: administrator
      operationId: "Baask8sGroupInspect"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Baask8s group identifier"
        required: true
        type: "integer"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Baask8sGroup"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "Baask8sGroup not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8sGroup not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    put:
      tags:
      - "baask8s_groups"
      summary: "Update an baask8s group"
      description: |
        Update an baask8s group.
        **Access policy**: administrator
      operationId: "Baask8sGroupUpdate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Baask8sGroup identifier"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Baask8sGroup details"
        required: true
        schema:
          $ref: "#/definitions/Baask8sGroupUpdateRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Baask8sGroup"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        404:
          description: "Baask8sGroup not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8sGroup not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
        503:
          description: "Baask8sGroup management disabled"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8sGroup management is disabled"
    delete:
      tags:
      - "baask8s_groups"
      summary: "Remove an baask8s group"
      description: |
        Remove an baask8s group.
        **Access policy**: administrator
      operationId: "Baask8sGroupDelete"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Baask8sGroup identifier"
        required: true
        type: "integer"
      responses:
        204:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "Baask8sGroup not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8sGroup not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
        503:
          description: "Baask8sGroup management disabled"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Baask8sGroup management is disabled"
...
  /users/{id}:
    get:
      tags:
      - "users"
      summary: "Inspect a user"
      description: |
        Retrieve details about a user.
        **Access policy**: administrator
      operationId: "UserInspect"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "User identifier"
        required: true
        type: "integer"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "User not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    put:
      tags:
      - "users"
      summary: "Update a user"
      description: |
        Update user details. A regular user account can only update his details.
        **Access policy**: authenticated
      operationId: "UserUpdate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "User identifier"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "User details"
        required: true
        schema:
          $ref: "#/definitions/UserUpdateRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "User not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    delete:
      tags:
      - "users"
      summary: "Remove a user"
      description: |
        Remove a user.
        **Access policy**: administrator
      operationId: "UserDelete"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "User identifier"
        required: true
        type: "integer"
      responses:
        204:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "User not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
  /team_memberships:
    get:
      tags:
      - "team_memberships"
      summary: "List team memberships"
      description: |
        List team memberships. Access is only available to administrators and team leaders.
        **Access policy**: restricted
      operationId: "TeamMembershipList"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TeamMembershipListResponse"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "team_memberships"
      summary: "Create a new team membership"
      description: |
        Create a new team memberships. Access is only available to administrators leaders of the associated team.
        **Access policy**: restricted
      operationId: "TeamMembershipCreate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - in: "body"
        name: "body"
        description: "Team membership details"
        required: true
        schema:
          $ref: "#/definitions/TeamMembershipCreateRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TeamMembership"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        409:
          description: "Team membership already exists"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Team membership already exists for this user and team."
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
  /templates:
    get:
      tags:
      - "templates"
      summary: "List available templates"
      description: |
        List available templates.
        Administrator templates will not be listed for non-administrator users.
        **Access policy**: restricted
      operationId: "TemplateList"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TemplateListResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "templates"
      summary: "Create a new template"
      description: |
        Create a new template.
        **Access policy**: administrator
      operationId: "TemplateCreate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - in: "body"
        name: "body"
        description: "Template details"
        required: true
        schema:
          $ref: "#/definitions/TemplateCreateRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Template"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  TeamMembership:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Membership identifier"
      UserID:
        type: "integer"
        example: 1
        description: "User identifier"
      TeamID:
        type: "integer"
        example: 1
        description: "Team identifier"
      Role:
        type: "integer"
        example: 1
        description: "Team role (1 for team leader and 2 for team member)"
...
  User:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "User identifier"
      Username:
        type: "string"
        example: "bob"
        description: "Username"
      Password:
        type: "string"
        example: "passwd"
        description: "Password"
      Role:
        type: "integer"
        example: 1
        description: "User role (1 for administrator account and 2 for regular account)"
...
  Pair:
    properties:
      name:
        type: "string"
        example: "name"
      value:
        type: "string"
        example: "value"
...
  Baask8sGroup:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Baask8s group identifier"
      Name:
        type: "string"
        example: "my-baask8s-group"
        description: "Baask8s group name"
      Description:
        type: "string"
        example: "Description associated to the baask8s group"
        description: "Baask8s group description"
      AuthorizedUsers:
        type: "array"
        description: "List of user identifiers authorized to connect to this baask8s group. Will be inherited by baask8ss that are part of the group"
        items:
          type: "integer"
          example: 1
          description: "User identifier"
      AuthorizedTeams:
        type: "array"
        description: "List of team identifiers authorized to connect to this baask8s. Will be inherited by baask8ss that are part of the group"
        items:
          type: "integer"
          example: 1
          description: "Team identifier"
      Labels:
        type: "array"
        items:
          $ref: "#/definitions/Pair"
...
  GenericError:
    type: "object"
    properties:
      err:
        type: "string"
        example: "Something bad happened"
        description: "Error message"
...
  Baask8sGroupUpdateRequest:
    type: "object"
    properties:
      Name:
        type: "string"
        example: "my-baask8s-group"
        description: "Baask8s group name"
      Description:
        type: "string"
        example: "Baask8s group description"
        description: "Baask8s group description"
      Labels:
        type: "array"
        items:
          $ref: "#/definitions/Pair"
      AssociatedBaask8ss:
        type: "array"
        description: "List of baask8s identifiers that will be part of this group"
        items:
          type: "integer"
          example: 1
          description: "Baask8s identifier"
...
  UserUpdateRequest:
    type: "object"
    properties:
      Password:
        type: "string"
        example: "cg9Wgky3"
        description: "Password"
      Role:
        type: "integer"
        example: 1
        description: "User role (1 for administrator account and 2 for regular account)"
...
  TeamMembershipCreateRequest:
    type: "object"
    required:
    - "UserID"
    - "TeamID"
    - "Role"
    properties:
      UserID:
        type: "integer"
        example: 1
        description: "User identifier"
      TeamID:
        type: "integer"
        example: 1
        description: "Team identifier"
      Role:
        type: "integer"
        example: 1
        description: "Role for the user inside the team (1 for leader and 2 for regular member)"
  TeamMembershipListResponse:
    type: "array"
    items:
      $ref: "#/definitions/TeamMembership"
...
  TemplateListResponse:
    type: "array"
    items:
      $ref: "#/definitions/Template"
  TemplateCreateRequest:
    type: "object"
    required:
    - "type"
    - "title"
    - "description"
    properties:
      type:
        type: "integer"
        example: 1
        description: "Template type. Valid values are: 1 (container), 2 (Swarm stack) or 3 (Compose stack)"
      title:
        type: "string"
        example: "Nginx"
        description: "Title of the template"
      description:
        type: "string"
        example: "High performance web server"
        description: "Description of the template"
      administrator_only:
        type: "boolean"
        example: true
        description: "Whether the template should be available to administrators only"
      image:
        type: "string"
        example: "nginx:latest"
        description: "Image associated to a container template. Mandatory for a container template"
      repository:
        $ref: "#/definitions/TemplateRepository"
      name:
        type: "string"
        example: "mystackname"
        description: "Default name for the stack/container to be used on deployment"
      logo:
        type: "string"
        example: "https://cloudinovasi.id/assets/img/logos/nginx.png"
        description: "URL of the template's logo"
      env:
        type: "array"
        description: "A list of environment variables used during the template deployment"
        items:
          $ref: "#/definitions/TemplateEnv"
      note:
        type: "string"
        example: "This is my <b>custom</b> template"
        description: "A note that will be displayed in the UI. Supports HTML content"
      platform:
        type: "string"
        example: "linux"
        description: "Platform associated to the template. Valid values are: 'linux', 'windows' or leave empty for multi-platform"
      categories:
        type: "array"
        description: "A list of categories associated to the template"
        items:
          type: "string"
          example: "database"
      registry:
        type: "string"
        example: "quay.io"
        description: "The URL of a registry associated to the image for a container template"
      command:
        type: "string"
        example: "ls -lah"
        description: "The command that will be executed in a container template"
      network:
        type: "string"
        example: "mynet"
        description: "Name of a network that will be used on container deployment if it exists inside the environment"
      volumes:
        type: "array"
        description: "A list of volumes used during the container template deployment"
        items:
          $ref: "#/definitions/TemplateVolume"
      ports:
        type: "array"
        description: "A list of ports exposed by the container"
        items:
          type: "string"
          example: "8080:80/tcp"
      labels:
        type: "array"
        description: "Container labels"
        items:
          $ref: '#/definitions/Pair'
      privileged:
        type: "boolean"
        example: true
        description: "Whether the container should be started in privileged mode"
      interactive:
        type: "boolean"
        example: true
        description: "Whether the container should be started in interactive mode (-i -t equivalent on the CLI)"
      restart_policy:
        type: "string"
        example: "on-failure"
        description: "Container restart policy"
      hostname:
        type: "string"
        example: "mycontainer"
        description: "Container hostname"
...
  Template:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
        description: "Template identifier"
      type:
        type: "integer"
        example: 1
        description: "Template type. Valid values are: 1 (container), 2 (Swarm stack) or 3 (Compose stack)"
      title:
        type: "string"
        example: "Nginx"
        description: "Title of the template"
      description:
        type: "string"
        example: "High performance web server"
        description: "Description of the template"
      administrator_only:
        type: "boolean"
        example: true
        description: "Whether the template should be available to administrators only"
      image:
        type: "string"
        example: "nginx:latest"
        description: "Image associated to a container template. Mandatory for a container template"
      repository:
        $ref: "#/definitions/TemplateRepository"
      name:
        type: "string"
        example: "mystackname"
        description: "Default name for the stack/container to be used on deployment"
      logo:
        type: "string"
        example: "https://cloudinovasi.id/assets/img/logos/nginx.png"
        description: "URL of the template's logo"
      env:
        type: "array"
        description: "A list of environment variables used during the template deployment"
        items:
          $ref: "#/definitions/TemplateEnv"
      note:
        type: "string"
        example: "This is my <b>custom</b> template"
        description: "A note that will be displayed in the UI. Supports HTML content"
      platform:
        type: "string"
        example: "linux"
        description: "Platform associated to the template. Valid values are: 'linux', 'windows' or leave empty for multi-platform"
      categories:
        type: "array"
        description: "A list of categories associated to the template"
        items:
          type: "string"
          example: "database"
      registry:
        type: "string"
        example: "quay.io"
        description: "The URL of a registry associated to the image for a container template"
      command:
        type: "string"
        example: "ls -lah"
        description: "The command that will be executed in a container template"
      network:
        type: "string"
        example: "mynet"
        description: "Name of a network that will be used on container deployment if it exists inside the environment"
      volumes:
        type: "array"
        description: "A list of volumes used during the container template deployment"
        items:
          $ref: "#/definitions/TemplateVolume"
      ports:
        type: "array"
        description: "A list of ports exposed by the container"
        items:
          type: "string"
          example: "8080:80/tcp"
      labels:
        type: "array"
        description: "Container labels"
        items:
          $ref: '#/definitions/Pair'
      privileged:
        type: "boolean"
        example: true
        description: "Whether the container should be started in privileged mode"
      interactive:
        type: "boolean"
        example: true
        description: "Whether the container should be started in interactive mode (-i -t equivalent on the CLI)"
      restart_policy:
        type: "string"
        example: "on-failure"
        description: "Container restart policy"
      hostname:
        type: "string"
        example: "mycontainer"
        description: "Container hostname"
  TemplateVolume:
    type: "object"
    properties:
      container:
        type: "string"
        example: "/data"
        description: "Path inside the container"
      bind:
        type: "string"
        example: "/tmp"
        description: "Path on the host"
      readonly:
        type: "boolean"
        example: true
        description: "Whether the volume used should be readonly"
  TemplateEnv:
    type: "object"
    properties:
      name:
        type: "string"
        example: "MYSQL_ROOT_PASSWORD"
        description: "name of the environment variable"
      label:
        type: "string"
        example: "Root password"
        description: "Text for the label that will be generated in the UI"
      description:
        type: "string"
        example: "MySQL root account password"
        description: "Content of the tooltip that will be generated in the UI"
      default:
        type: "string"
        example: "default_value"
        description: "Default value that will be set for the variable"
      preset:
        type: "boolean"
        example: true
        description: "If set to true, will not generate any input for this variable in the UI"
      select:
        type: "array"
        description: "A list of name/value that will be used to generate a dropdown in the UI"
        items:
          $ref: '#/definitions/TemplateEnvSelect'
  TemplateEnvSelect:
    type: "object"
    properties:
      text:
        type: "string"
        example: "text value"
        description: "Some text that will displayed as a choice"
      value:
        type: "string"
        example: "value"
        description: "A value that will be associated to the choice"
      default:
        type: "boolean"
        example: true
        description: "Will set this choice as the default choice"
  TemplateRepository:
    type: "object"
    required:
    - "URL"
    properties:
      URL:
        type: "string"
        example: "https://github.com/baasapi/baasapi-compose"
        description: "URL of a git repository used to deploy a stack template. Mandatory for a Swarm/Compose stack template"
      stackfile:
        type: "string"
        example: "./subfolder/docker-compose.yml"
        description: "Path to the stack file inside the git repository"
...